+++ poly-commitment/src/kzg.rs
@@ -16,10 +16,13 @@ use crate::{
         combine_commitments, BatchEvaluationProof, BlindedCommitment, CommitmentCurve, Evaluation,
         PolyComm,
     },
+    error::CommitmentError,
     ipa::SRS,
     utils::combine_polys,
-    CommitmentError, PolynomialsToCombine, SRS as SRSTrait,
+    PolynomialsToCombine, SRS as SRSTrait,
 };
+use alloc::vec;
+use alloc::vec::Vec;
 
 use ark_ec::{pairing::Pairing, AffineRepr, VariableBaseMSM};
 use ark_ff::{One, PrimeField, Zero};
@@ -27,12 +30,13 @@ use ark_poly::{
     univariate::{DenseOrSparsePolynomial, DensePolynomial},
     DenseUVPolynomial, EvaluationDomain, Evaluations, Polynomial, Radix2EvaluationDomain as D,
 };
+use core::ops::Neg;
 use mina_poseidon::FqSponge;
+#[cfg(feature = "std")]
 use rand::thread_rng;
 use rand_core::{CryptoRng, RngCore};
 use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
-use std::ops::Neg;
 
 /// Combine the (chunked) evaluations of multiple polynomials.
 ///
@@ -135,6 +139,7 @@ pub struct PairingSRS<Pair: Pairing> {
     pub verifier_srs: SRS<Pair::G2Affine>,
 }
 
+#[cfg(feature = "std")]
 impl<
         F: PrimeField,
         G: CommitmentCurve<ScalarField = F>,
@@ -176,6 +181,7 @@ impl<Pair: Pairing> Clone for PairingSRS<Pair> {
     }
 }
 
+#[cfg(feature = "std")]
 impl<
         F: PrimeField,
         G: CommitmentCurve<ScalarField = F>,
@@ -239,6 +245,7 @@ impl<
     }
 }
 
+#[cfg(feature = "std")]
 impl<
         F: PrimeField,
         G: CommitmentCurve<ScalarField = F>,
@@ -387,6 +394,7 @@ fn divisor_polynomial<F: PrimeField>(elm: &[F]) -> DensePolynomial<F> {
         .unwrap()
 }
 
+#[cfg(feature = "std")]
 impl<
         F: PrimeField,
         G: CommitmentCurve<ScalarField = F>,